avro sink support

+++ cdc/model/sink.go
@@ -74,7 +74,7 @@ type RowChangedEvent struct {
 
 	SchemaID int64 `json:"schema-id,omitempty"`
 
-	SchemaVersion uint16 `json:"schema-version,omitempty"`
+	SchemaVersion uint64 `json:"schema-version,omitempty"`
 
 	// if the table of this row only has one unique index(includes primary key),
 	// IndieMarkCol will be set to the name of the unique index
